+++ ScrumEnv/app/config/parameters.yml.dist~
@@ -4,9 +4,9 @@
 parameters:
     database_host:     127.0.0.1
     database_port:     ~
-    database_name:     ScrumDatabase
-    database_user:     ScrumUser
-    database_password: null
+    database_name:     symfony
+    database_user:     ~
+    database_password: ~
     # You should uncomment this if you want use pdo_sqlite
     # database_path: "%kernel.root_dir%/data.db3"
