+++ ScrumEnv/composer.json
@@ -55,7 +55,8 @@
             "file": "app/config/parameters.yml"
         },
         "branch-alias": {
-            "dev-master": "2.7-dev"
+            "dev-master": "2.7-dev",
+            "keep-outdated":true
         }
     }
 }
