--- ScrumEnv/src/AppBundle/Controller/backlogController.php~
@@ -1,25 +1,0 @@
-<?php>
-
-namespace AppBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Component\HttpFoundation\Response;
-
-class BacklogController
-{
-    /*
-    *@Route("/{user}/{project}/backlog")
-    */
-    public function displayBacklog()
-    {
-        //phony data for testing purpose
-        $usList = array(
-            "1" => array("id" => 1,"description" => "this is a user story", "cost" => 2, "priority" => 1),
-            "2" => array("id" => 2,"description" => "this is an other user story for my project", "cost" => 3, "priority" => 2));
-        return this->render("backlog.html.twig", usList);
-    }
-
-}
-
-
