--- ScrumEnv/src/AppBundle/Entity/US.php
@@ -1,188 +1,0 @@
-<?php
-
-namespace AppBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
-* @ORM\Entity
-* @ORM\Table(name="US")
-*/
-class US
-{    
-    /**
-    * @ORM\Column(type="integer")
-    * @ORM\Id
-    * @ORM\GeneratedValue(strategy="AUTO")
-    */
-    protected $_id;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $owner;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $project;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $id;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $description;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $cost;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $priority;
-    
-
-    /**
-     * Get id
-     *
-     * @return integer
-     */
-    public function getId()
-    {
-        return $this->_id;
-    }
-
-    /**
-     * Set owner
-     *
-     * @param string $owner
-     *
-     * @return US
-     */
-    public function setOwner($owner)
-    {
-        $this->owner = $owner;
-
-        return $this;
-    }
-
-    /**
-     * Get owner
-     *
-     * @return string
-     */
-    public function getOwner()
-    {
-        return $this->owner;
-    }
-
-    /**
-     * Set project
-     *
-     * @param string $project
-     *
-     * @return US
-     */
-    public function setProject($project)
-    {
-        $this->project = $project;
-
-        return $this;
-    }
-
-    /**
-     * Get project
-     *
-     * @return string
-     */
-    public function getProject()
-    {
-        return $this->project;
-    }
-
-    /**
-     * Set id
-     *
-     * @param string $id
-     *
-     * @return US
-     */
-    public function setId($id)
-    {
-        $this->id = $id;
-
-        return $this;
-    }
-
-    /**
-     * Set description
-     *
-     * @param string $description
-     *
-     * @return US
-     */
-    public function setDescription($description)
-    {
-        $this->description = $description;
-
-        return $this;
-    }
-
-    /**
-     * Get description
-     *
-     * @return string
-     */
-    public function getDescription()
-    {
-        return $this->description;
-    }
-
-    /**
-     * Set cost
-     *
-     * @param string $cost
-     *
-     * @return US
-     */
-    public function setCost($cost)
-    {
-        $this->cost = $cost;
-
-        return $this;
-    }
-
-    /**
-     * Get cost
-     *
-     * @return string
-     */
-    public function getCost()
-    {
-        return $this->cost;
-    }
-
-    /**
-     * Set priority
-     *
-     * @param string $priority
-     *
-     * @return US
-     */
-    public function setPriority($priority)
-    {
-        $this->priority = $priority;
-
-        return $this;
-    }
-
-    /**
-     * Get priority
-     *
-     * @return string
-     */
-    public function getPriority()
-    {
-        return $this->priority;
-    }
-}
