--- ScrumEnv/src/AppBundle/Entity/US.php~
@@ -1,44 +1,0 @@
-<?php
-
-namespace AppBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
-* @ORM\Entity
-* @ORM\Table(name="US")
-*/
-class US
-{    
-    /**
-    * @ORM\Column(type="integer")
-    * @ORM\Id
-    * @ORM\GeneratedValue(strategy="AUTO")
-    */
-    protected $_id;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $owner;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $project;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $id;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $description;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $cost;
-    /**
-    * @ORM\Column(type="string", length=100)
-    */
-    public $priority;
-    
-}
