+++ ScrumEnv/src/ProjectBundle/ProjectBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ProjectBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class ProjectBundle extends Bundle
+{
+}
