--- ScrumEnv/src/test/ProjectBundle/Resources/views/Default/index.html.twig
@@ -1,20 +1,0 @@
-
-<html>
-<head>
-	<link href="{{ asset('bundles/testProject/css/main.css') }}" rel="stylesheet"  type="text/css"/>
-</head>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-	<body >
-	<div id="Project">
-			<p id="TitlePage">Welcome on website</p>
-			<div>
-				<a href="{{ path('go_to_backlog') }}">Create a Backlog</a>
-			</br></br>
-				 <a href="{{ path('go_to_sprintIndex', { 'name':'guillaume' }) }}"> Sprints page </a>
-			</div>
-	</div>
-
-	
-	</body>
-</html>
-
