--- ScrumEnv/src/test/ProjectBundle/Resources/views/Kanban/Sprint_page.html.twig
@@ -1,18 +1,0 @@
-
-<html>
-<head>
-	<link href="{{ asset('bundles/testProject/css/main.css') }}" rel="stylesheet"  type="text/css"/>
-</head>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-	<body >
-	<div id="Project">
-			<p id="TitlePage">Welcome on sprints page</p>
-			<div>
-				<a href="{{ path('go_to_Kanban')}}"> Create kanban for sprint </a>
-			</div>
-	</div>
-
-	
-	</body>
-</html>
-
