--- ScrumEnv/src/test/ProjectBundle/Tests/Controller/DefaultControllerTest.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace test\ProjectBundle\Tests\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class DefaultControllerTest extends WebTestCase
-{
-    public function testIndex()
-    {
-        $client = static::createClient();
-
-        $crawler = $client->request('GET', '/hello/Fabien');
-
-        $this->assertTrue($crawler->filter('html:contains("Hello Fabien")')->count() > 0);
-    }
-
-}
